application helpers extracted into a module
